fix(p4): rare scenario of file version conflict
